requirements.txt
django
djangorestframework
django-cors-headers
faker
factory-boy